requirements.txt
flask>=3.0.0
requests>=2.31.0
gunicorn>=21.2.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
